+++ NEWS.md
@@ -4,3 +4,6 @@
 - Updated TL entry probabilities and choice proportions in `subopt_full` for Cunningham & Shahan 2019 (newly digitized values).
 - Adding versioning to `subopt_full` and `subopt_avian` data sets.
 - Improved "Signal Discrimination" notes in `choice_params()` documentation.
+
+## 2025-07-22
+- Updated `choice_params()` to only display recycling info message if 3 or more arguments differ in length.